Import initial content from the old website
Just the default index page for now, with its corresponding support
files (a few CSS files and icons that are mentioned).

Removed the reference to the index.en.html alternate, as that is not
there.

Updated the webmaster email to the cod.ifi.cat domain, because I am
not sure the one for turro.net works (it should! but I have not
checked in a while)